Split widget file into two
